Move `list_of()` utils to `type-list-of.c`
